refactor(MIGR-05): share one Kea reconciliation plan between dry-run and import
kea_migration_api.py (dry-run preview) and importer.py (real import) each
independently recomputed the same sequence — team merge, OpenFGA tuple
transform, relation-subject resolution, team_member derivation, admin-less
detection, platform-role-grant resolution — using their own separately
constructed KeaUserResolver. Any change to that sequencing had to be made
in both places by hand, and a dry-run followed by a real import triggered
two full bulk Keycloak sweeps instead of one.

Extract build_kea_reconciliation_plan() into kea_reconciliation.py: a pure,
zero-write function that both callers now execute identically, sharing the
same resolver and KeaReconciliationReport instance passed in by the caller.

Moving _merge_kea_team_rows/_realm_group_names/_realm_platform_role_grants/
transform_kea_tuples (and their private helpers) from importer.py into
kea_reconciliation.py along the way also drops kea_migration_api.py's cross
-import of importer.py entirely, and shrinks the amount of kea-only code
living inside importer.py (the permanent executor) from ~360 lines to zero
-- it now lives solely in the two files already tagged for deletion after
the S3NS cutover.

No behavior change: KeaDryRunResponse's contract, resolution semantics
(MATCHED/RELINKED/PENDING), and write-ordering are unchanged. Verified
against the existing test suite (kea bundle/reconciliation/users/warnings-
cap/manifest/team-metadata/task-observability, 76 tests) plus ruff, ruff
format, and basedpyright, all clean.
