+++ apps/control-plane-backend/control_plane_backend/import_export/importer.py
@@ -75,7 +75,6 @@ from __future__ import annotations
 
 import asyncio
 import logging
-import re
 from collections.abc import Awaitable, Callable
 from dataclasses import dataclass, field
 from datetime import datetime, timezone
@@ -133,17 +132,10 @@ from control_plane_backend.import_export.bundle import KBundle
 from control_plane_backend.import_export.kea_reconciliation import (  # KEA CUTOVER 2026 — delete with kea_reconciliation.py
     KeaReconciliationReport,
     KeaUserResolver,
-    derive_team_member_relations,
-    drop_orphan_team_relations,
-    drop_orphan_teams,
-    find_admin_less_teams,
-    format_usernames_for_warning,
-    kea_known_group_ids,
+    build_kea_reconciliation_plan,
     kea_username_by_sub,
     resolve_agent_team_index,
     resolve_creator_index,
-    resolve_platform_grants,
-    resolve_relation_subjects,
     resolve_resource_authors,
     resolve_tag_owner_ids,
 )
@@ -464,276 +456,6 @@ def _strip_front_matter(content: str) -> str:
     """
     _, sep, body = content.partition("\n---\n")
     return body.strip() if sep else content.strip()
-
-
-# ── Kea teams & platform roles from the Keycloak realm export ─────────────────
-
-
-def _realm_group_names(realm: dict[str, Any] | None) -> dict[str, str]:
-    """Return group_id → group name from a Keycloak realm export.
-
-    On kea a team's id IS its Keycloak group id, and the group name is the
-    only place the team's name exists (kea `teammetadata` has no name column).
-    Walks sub-groups too, defensively — kea teams are top-level groups.
-    """
-    names: dict[str, str] = {}
-
-    def _walk(groups: list[dict[str, Any]]) -> None:
-        for group in groups:
-            group_id = group.get("id")
-            name = group.get("name")
-            if isinstance(group_id, str) and isinstance(name, str) and name:
-                names[group_id] = name
-            _walk(group.get("subGroups") or [])
-
-    _walk((realm or {}).get("groups") or [])
-    return names
-
-
-def _merge_kea_team_rows(
-    raw_team_metadata: list[dict[str, Any]],
-    tuples: list[dict[str, Any]],
-    group_names: dict[str, str],
-    report: MigrationReport,
-) -> list[dict[str, Any]]:
-    """Build the full kea team list: every team referenced by tuples, named.
-
-    Kea materialises a `teammetadata` row only when a team was customized
-    (description / privacy / banner) — an untouched team exists solely as a
-    Keycloak group plus OpenFGA tuples. Swift requires a `teammetadata` row
-    (NOT NULL unique `name`) for every team, so this merges:
-    - team ids ← every `team:<id>` reference in the tuple dump (excluding the
-      kea shared personal team and swift-style personal ids),
-    - customization ← the kea `teammetadata` row when present,
-    - name ← the realm export's group name; falls back to the id (warned)
-      when the bundle has no realm export or the group is gone.
-    """
-    teams: dict[str, dict[str, Any]] = {}
-    for t in tuples:
-        for ref in (t.get("object", ""), t.get("user", "")):
-            if not ref.startswith("team:"):
-                continue
-            team_id = ref.removeprefix("team:")
-            if team_id == _KEA_SHARED_PERSONAL_TEAM_ID or team_id.startswith(
-                "personal-"
-            ):
-                continue
-            teams.setdefault(team_id, {"id": team_id})
-    for row in raw_team_metadata:
-        team_id = row.get("id")
-        if not team_id:
-            continue
-        merged = teams.setdefault(team_id, {})
-        merged.update(row)
-        merged["id"] = team_id
-
-    unnamed: list[str] = []
-    for team_id, row in teams.items():
-        realm_name = group_names.get(team_id)
-        if realm_name:
-            row["name"] = realm_name
-        elif not row.get("name"):
-            # `_import_team_metadata` falls back to name=id; surface it.
-            unnamed.append(team_id)
-    if unnamed:
-        report.warnings.append(
-            f"{len(unnamed)} team(s) have no name in the bundle (keycloak "
-            "realm export missing or group deleted) — they are named by "
-            f"their id: {', '.join(sorted(unnamed))}"
-        )
-    return [teams[team_id] for team_id in sorted(teams)]
-
-
-def _realm_platform_role_grants(
-    realm: dict[str, Any] | None,
-) -> tuple[list[tuple[str, RelationType]], list[str]]:
-    """Derive swift platform-role grants from a realm export's users, if any.
-
-    Kea platform roles are Keycloak realm roles (`admin`/`editor`/`viewer`)
-    carried per-user — present only in a FULL realm export (`kc export
-    --users`); a partial-export has no `users[]` and yields nothing here.
-    Mapping (AUTHZ target model): `admin` → platform_admin, `viewer` →
-    platform_observer, `editor` → dropped (no swift equivalent — returned
-    separately for a warning).
-    """
-    grants: list[tuple[str, RelationType]] = []
-    dropped_editors: list[str] = []
-    for user in (realm or {}).get("users") or []:
-        sub = user.get("id")
-        if not isinstance(sub, str) or not sub:
-            continue
-        roles = set(user.get("realmRoles") or [])
-        if "admin" in roles:
-            grants.append((sub, RelationType.PLATFORM_ADMIN))
-        if "viewer" in roles:
-            grants.append((sub, RelationType.PLATFORM_OBSERVER))
-        if "editor" in roles:
-            dropped_editors.append(user.get("username") or sub)
-    return grants, dropped_editors
-
-
-# ── Kea → swift OpenFGA tuple transformation (MIGR-05.04) ─────────────────────
-
-_UUID_RE = re.compile(
-    r"^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-"
-    r"[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
-)
-
-# Kea team roles are hierarchical (owner ⊃ manager ⊃ member); swift team_admin
-# and team_editor are orthogonal (REBAC.md "hard cross-write rule"), so a kea
-# owner must receive BOTH to keep the content authority it had. Mapping
-# approved by the developer on 2026-07-24. `team_analyst` has no kea source
-# and is never synthesized.
-_KEA_TEAM_ROLE_TO_SWIFT: dict[str, tuple[RelationType, ...]] = {
-    "owner": (RelationType.TEAM_ADMIN, RelationType.TEAM_EDITOR),
-    "manager": (RelationType.TEAM_EDITOR,),
-    "member": (RelationType.TEAM_MEMBER,),
-}
-
-# The kea shared personal-space team id (`PERSONAL_TEAM_ID` on main). Swift
-# personal spaces are per-user (`personal-{uid}`) and their access tuple is
-# self-healed on first use, so kea personal-team tuples must be dropped, not
-# translated.
-_KEA_SHARED_PERSONAL_TEAM_ID = "personal"
-
-_RESOURCE_BY_PREFIX: dict[str, Resource] = {
-    "user": Resource.USER,
-    "team": Resource.TEAM,
-    "organization": Resource.ORGANIZATION,
-    "agent": Resource.AGENT,
-    "tag": Resource.TAGS,
-    "document": Resource.DOCUMENTS,
-}
-
-
-@dataclass
-class KeaTupleTransform:
-    """Outcome of transforming a kea tuple dump into swift relations."""
-
-    relations: list[Relation] = field(default_factory=list)
-    dropped_personal: int = 0
-    dropped_non_uuid: int = 0
-    dropped_resource_parent: int = 0
-    dropped_unknown: int = 0
-    unknown_shapes: set[str] = field(default_factory=set)
-
-    @property
-    def dropped_total(self) -> int:
-        return (
-            self.dropped_personal
-            + self.dropped_non_uuid
-            + self.dropped_resource_parent
-            + self.dropped_unknown
-        )
-
-
-def transform_kea_tuples(tuples: list[dict[str, Any]]) -> KeaTupleTransform:
-    """Map a kea OpenFGA tuple dump onto the swift authorization model.
-
-    - team `owner`/`manager`/`member` tuples are aggregated per (user, team)
-      and re-emitted as swift `team_*` relations; a user with an elevated role
-      never also gets a redundant direct `team_member` tuple (schema.fga
-      derives it).
-    - `agent`/`tag`/`document` ownership and `team#organization`/`team#public`
-      tuples replay 1:1 (identical relation names on both models).
-    - dropped: anything touching the kea shared personal team, `resource#parent`
-      (resources migrate to prompt rows, which have no OpenFGA object),
-      non-UUID user subjects (pre-MIGR-04 username tuples), and any shape the
-      swift model does not know.
-    """
-    out = KeaTupleTransform()
-    team_roles: dict[tuple[str, str], set[str]] = {}
-    seen: set[tuple[str, str, str, str, str]] = set()
-
-    def _emit(
-        subject: RebacReference, relation: RelationType, resource: RebacReference
-    ) -> None:
-        key = (
-            subject.type.value,
-            str(subject.id),
-            relation.value,
-            resource.type.value,
-            str(resource.id),
-        )
-        if key in seen:
-            return
-        seen.add(key)
-        out.relations.append(
-            Relation(subject=subject, relation=relation, resource=resource)
-        )
-
-    def _split(ref: str) -> tuple[str, str]:
-        prefix, _, ident = ref.partition(":")
-        return prefix, ident
-
-    for t in tuples:
-        subj_type, subj_id = _split(t.get("user", ""))
-        rel = t.get("relation", "")
-        obj_type, obj_id = _split(t.get("object", ""))
-
-        touches_shared_personal = (
-            obj_type == "team" and obj_id == _KEA_SHARED_PERSONAL_TEAM_ID
-        ) or (subj_type == "team" and subj_id == _KEA_SHARED_PERSONAL_TEAM_ID)
-        if touches_shared_personal:
-            out.dropped_personal += 1
-            continue
-        if obj_type == "resource":
-            out.dropped_resource_parent += 1
-            continue
-        if subj_type == "user" and subj_id != "*" and not _UUID_RE.match(subj_id):
-            out.dropped_non_uuid += 1
-            continue
-
-        if (
-            obj_type == "team"
-            and subj_type == "user"
-            and rel in _KEA_TEAM_ROLE_TO_SWIFT
-        ):
-            team_roles.setdefault((subj_id, obj_id), set()).add(rel)
-            continue
-
-        replayable = (
-            (obj_type == "agent" and rel == "owner" and subj_type in ("user", "team"))
-            or (
-                obj_type == "tag"
-                and rel in ("owner", "editor", "viewer")
-                and subj_type in ("user", "team")
-            )
-            or (obj_type == "tag" and rel == "parent" and subj_type == "tag")
-            or (obj_type == "document" and rel == "parent" and subj_type == "tag")
-            or (
-                obj_type == "team"
-                and rel == "organization"
-                and subj_type == "organization"
-            )
-            or (obj_type == "team" and rel == "public" and subj_type == "user")
-        )
-        subject_res = _RESOURCE_BY_PREFIX.get(subj_type)
-        object_res = _RESOURCE_BY_PREFIX.get(obj_type)
-        if not replayable or subject_res is None or object_res is None:
-            out.dropped_unknown += 1
-            out.unknown_shapes.add(f"{subj_type} {rel} {obj_type}")
-            continue
-        _emit(
-            RebacReference(subject_res, subj_id),
-            RelationType(rel),
-            RebacReference(object_res, obj_id),
-        )
-
-    for (uid, team_id), kea_roles in sorted(team_roles.items()):
-        targets: set[RelationType] = set()
-        for role in kea_roles & {"owner", "manager"}:
-            targets.update(_KEA_TEAM_ROLE_TO_SWIFT[role])
-        if not targets and "member" in kea_roles:
-            targets.add(RelationType.TEAM_MEMBER)
-        for relation in sorted(targets, key=lambda r: r.value):
-            _emit(
-                RebacReference(Resource.USER, uid),
-                relation,
-                RebacReference(Resource.TEAM, team_id),
-            )
-
-    return out
 
 
 _STEP_LABELS: dict[str, str] = {
@@ -1601,111 +1323,28 @@ async def _run_import_body(
     raw_metadata = list(bundle.iter_table("metadata"))
     # Per-team branding + retention (CTRLP-12). Kea's `teammetadata` rows carry
     # `is_private` instead of `joining_mode` and may lack `name` — both handled
-    # by `_import_team_metadata`'s fallbacks below.
-    raw_team_metadata = list(bundle.iter_table(team_table))
-    keycloak_realm = None if is_swift_native else bundle.keycloak_realm()
-    if not is_swift_native:
-        # Kea only materialises a teammetadata row for customized teams; an
-        # untouched team exists solely as a Keycloak group + tuples. Build the
-        # complete team list (every tuple-referenced team), named from the
-        # realm export's groups — the only place kea stores team names.
-        raw_team_metadata = _merge_kea_team_rows(
-            raw_team_metadata,
-            tuples,
-            _realm_group_names(keycloak_realm),
-            report,
-        )
-        # KEA CUTOVER 2026 — drop teams referenced only by a stray/stale OpenFGA
-        # tuple with no matching Keycloak group (see kea_reconciliation.py). Without
-        # this, `_merge_kea_team_rows`'s own id-as-name fallback above would create a
-        # real, garbage-named team for a leftover tuple from an unrelated old export.
-        raw_team_metadata = drop_orphan_teams(
-            raw_team_metadata, kea_known_group_ids(keycloak_realm), kea_report
-        )
+    # by `_import_team_metadata`'s fallbacks below. Swift-native bundles use
+    # their raw `team_metadata` rows as-is; the kea path replaces this with
+    # `plan.team_metadata` below (merged + orphan-dropped).
+    raw_team_metadata = list(bundle.iter_table(team_table)) if is_swift_native else []
 
-    # ── KEA CUTOVER 2026 — resolve every Keycloak-dependent write BEFORE the
-    # Postgres transaction opens: relation-subject resolution, team_member
-    # derivation, and platform-role-grant resolution all need the (bulk-
-    # prefetched, see kea_reconciliation.py) `kea_resolver` — none of them
-    # depend on anything the transaction below writes (team ids are bundle-
-    # derived Keycloak group ids, never Postgres-generated). This is a
-    # fail-fast improvement, not a correctness fix — the writes these produce
-    # were already safely re-runnable (idempotent skip-if-exists Postgres
-    # inserts, idempotent OpenFGA `add_relation`) even when resolved after the
-    # transaction. Doing it first means a Keycloak failure aborts the whole
-    # import before any Postgres row is written this run, instead of after
-    # agents/tags/resources/metadata/team_metadata already committed.
+    # ── KEA CUTOVER 2026 — team merge, OpenFGA tuple restore, and platform-role
+    # resolution, all BEFORE the Postgres transaction opens: `build_kea_reconciliation_plan`
+    # is the single source of truth shared with the dry-run preview
+    # (`kea_migration_api.py`) — same sequencing, same (bulk-prefetched)
+    # `kea_resolver`, no independent recomputation. None of this depends on
+    # anything the transaction below writes (team ids are bundle-derived
+    # Keycloak group ids, never Postgres-generated), so a Keycloak failure
+    # aborts the whole import before any Postgres row is written this run.
     resolved_relations: list[Relation] = []
     realm_grants: list[tuple[str, RelationType]] = []
-    if not is_swift_native and tuples:
-        transform = transform_kea_tuples(tuples)
-        report.tuples_dropped = transform.dropped_total
-        # Mirror `drop_orphan_teams` on the tuple-restore path: a team excluded
-        # from `teammetadata` above must not still get its OpenFGA tuples
-        # written (see `drop_orphan_team_relations`' docstring).
-        transform.relations = drop_orphan_team_relations(
-            transform.relations, kea_report.orphan_teams_dropped
-        )
-        if transform.dropped_non_uuid:
-            report.warnings.append(
-                f"{transform.dropped_non_uuid} OpenFGA tuple(s) with a "
-                "non-UUID user subject dropped (pre-MIGR-04 username tuples)"
-            )
-        if transform.dropped_unknown:
-            report.warnings.append(
-                f"{transform.dropped_unknown} OpenFGA tuple(s) dropped — no "
-                "swift equivalent for: " + ", ".join(sorted(transform.unknown_shapes))
-            )
-
-        # 1) rewrite every USER-typed subject to its resolved swift sub
-        #    (dropping ones that can't be resolved yet — see
-        #    kea_reconciliation.py); 2) derive `team_member` grants from
-        #    kea's Keycloak group membership, the one Fred relation with no
-        #    OpenFGA-tuple source on kea.
-        resolved_relations = await resolve_relation_subjects(
-            transform.relations, kea_username_index, kea_resolver, kea_report
-        )
-        already_elevated = {
-            (str(r.subject.id), str(r.resource.id))
-            for r in resolved_relations
-            if r.relation
-            in (
-                RelationType.TEAM_ADMIN,
-                RelationType.TEAM_EDITOR,
-                RelationType.TEAM_ANALYST,
-            )
-        }
-        group_name_to_team_id = {
-            row["name"]: row["id"] for row in raw_team_metadata if row.get("name")
-        }
-        member_relations = await derive_team_member_relations(
-            kea_realm_for_reconciliation,
-            group_name_to_team_id,
-            already_elevated,
-            kea_username_index,
-            kea_resolver,
-            kea_report,
-        )
-        resolved_relations = resolved_relations + member_relations
-        kea_report.admin_less_teams = find_admin_less_teams(
-            {row["id"] for row in raw_team_metadata}, resolved_relations
-        )
-
     if not is_swift_native:
-        # Kea platform roles are Keycloak realm roles per user — never
-        # tuples. They only travel in a FULL realm export (`kc export
-        # --users`); a partial-export yields nothing here and the grants
-        # then come from `users.json` (or manual bootstrap).
-        realm_grants, dropped_editors = _realm_platform_role_grants(keycloak_realm)
-        if dropped_editors:
-            report.warnings.append(
-                f"kea platform role 'editor' dropped for {len(dropped_editors)} "
-                f"user(s) (no swift equivalent): "
-                f"{format_usernames_for_warning(dropped_editors)}"
-            )
-        realm_grants = await resolve_platform_grants(
-            realm_grants, kea_username_index, kea_resolver, kea_report
-        )
+        plan = await build_kea_reconciliation_plan(bundle, kea_resolver, kea_report)
+        raw_team_metadata = plan.team_metadata
+        resolved_relations = plan.resolved_relations
+        realm_grants = plan.realm_platform_grants
+        report.tuples_dropped = plan.tuples_dropped_total
+        report.warnings.extend(plan.warnings)
     # ── END KEA CUTOVER 2026 (pre-transaction resolution) ──────────────────────
 
     # ── Phases 2–4: all writes in a single atomic transaction ─────────────────
